Adapt differet float precision

+++ src/core/pv_mod.F90
@@ -244,7 +244,7 @@ contains
 
     type(block_type), intent(in) :: block
     type(state_type), intent(inout) :: state
-    real(r8), intent(in) :: dt
+    real(8), intent(in) :: dt
 
     type(mesh_type), pointer :: mesh
     real(r8) u, v
